Rename honk bench folder more accurately
