+++ barretenberg/cpp/src/barretenberg/benchmark/CMakeLists.txt
@@ -1,6 +1,6 @@
 add_subdirectory(decrypt_bench)
 add_subdirectory(pippenger_bench)
 add_subdirectory(plonk_bench)
-add_subdirectory(honk_bench)
+add_subdirectory(ultra_bench)
 add_subdirectory(relations_bench)
 add_subdirectory(widgets_bench)
