+++ barretenberg/cpp/src/barretenberg/benchmark/ultra_bench/goblin.bench.cpp
@@ -1,7 +1,7 @@
 
 #include <benchmark/benchmark.h>
 
-#include "barretenberg/benchmark/honk_bench/benchmark_utilities.hpp"
+#include "barretenberg/benchmark/ultra_bench/benchmark_utilities.hpp"
 #include "barretenberg/goblin/goblin.hpp"
 #include "barretenberg/goblin/mock_circuits.hpp"
 #include "barretenberg/proof_system/circuit_builder/ultra_circuit_builder.hpp"
