+++ barretenberg/cpp/src/barretenberg/benchmark/ultra_bench/standard_plonk.bench.cpp
@@ -1,4 +1,4 @@
-#include "barretenberg/benchmark/honk_bench/benchmark_utilities.hpp"
+#include "barretenberg/benchmark/ultra_bench/benchmark_utilities.hpp"
 #include "barretenberg/plonk/composer/standard_composer.hpp"
 #include "barretenberg/proof_system/circuit_builder/standard_circuit_builder.hpp"
